Update .NET Core samples search libgdiplus library
These C# samples were updated to search for a secondary known location
for the libgdiplus shared library. This library is required to use .NET
Core with Linux.

+++ DotNETCore/Sample_Source/Images/ImageEmbedICCProfile/ImageEmbedICCProfile.cs
@@ -88,16 +88,15 @@ namespace ImageEmbedICCProfile
         static void Main(string[] args)
         {
             if (System.Runtime.InteropServices.RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform
-                    .OSX) &&
-                !File.Exists("/usr/local/lib/libgdiplus.dylib"))
+                .OSX) && !File.Exists("/usr/local/lib/libgdiplus.dylib"))
             {
                 Console.WriteLine("Please install libgdiplus first to access the System.Drawing namespace on macOS.");
                 return;
             }
 
             if (System.Runtime.InteropServices.RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform
-                    .Linux) &&
-                !File.Exists("/usr/lib64/libgdiplus.so"))
+                .Linux) && !System.IO.File.Exists("/usr/lib64/libgdiplus.so") &&
+                !System.IO.File.Exists("/usr/lib/libgdiplus.so"))
             {
                 Console.WriteLine("Please install libgdiplus first to access the System.Drawing namespace on Linux.");
                 return;
